Implemented systemd socket activation
This adds support for systemd socket activation which can improve boot
times while avoiding spurious "connection refused" errors in clients.
This uses the new `systemd_socket` crate (created by me).

It is a breaking change in lib API because the listener must be passed
separately since it's not reusable. The configuration is
backwards-compatible.