Use newttype for password
Apart from the fact that newtypes are just better design especially
around security, this implements validation of password according to
RFC5234 - rejects non-ASCII characters and control characters. This will
help avoid surprises later. We take this opportunity to modify Debug
impl to not print the password, just in case, so we don't leak it
accidentally.