Removed derive_more
The only actually used trait derived by `derive_more` was `Deref` in two
instances, which seemed a bit too much for an entire crate, so it was
removed and replaced with manual impl.

The referred cases actually resemble the `Deref` pseudo-inheritence
anti-pattern so maybe they should be rewritten but that's a different
topic.